move & rename: events에서 사용하는 상수 파일명 변경 및 디렉토리 이동
